Documentation and Code changes
1) RAG UI updated, the UI Elements refresh rctified
2) Documentation added
